pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastmcp-bigquery-server"
version = "0.1.0"
description = "FastMCP server for querying Google BigQuery with AWS-managed credentials"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "FastMCP Solutions" }
]
dependencies = [
  "fastmcp>=2.13.2",
  "google-cloud-bigquery>=3.25.0",
  "google-auth>=2.34.0",
  "boto3>=1.35.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
